os: fix how to download cork binary file
The existing cork binary provided by upstream, could not be used by
Flatcar SDK, because it checked for coreos-specific variables.
So let's use the cork v0.11.2, built from flatcar-master.

For now there's no GPG key for the cork binary.
Still we need to figure out how to generate the GPG key to be uploaded
to the cork release v0.11.2.