.github: move parts into docs.flatcar-linux.org
To be able to create pull requests on the site of
docs.flatcar-linux.org, we need to move the actual syncing parts
into docs.flatcar-linux.org.
Instead, on the side of flatcar-linux/docs, we need to send a repository
dispatch event, whenever a commit was pushed to the master branch.