vmware: Update and expand the documentation
Important basic information was missing, like the usage of ovftool
and how static IPs are configured.
The relation of Ignition and coreos-cloudinit is explained because
both use the same data source and Ignition has serveral limitations.
Depending on the use case, some important behavior was documented
such as which guestinfo variables result in what networkd setting
when coreos-cloudinit is used, or how custom templating should be
done with /run/metadata/coreos.