update-from-container-linux: Migrate /etc/coreos/update.conf
Any setting in /etc/coreos/update.conf was lost because the file
wasn't migrated to the new location.
Move it to the new location but keep a symlink. The SERVER parameter
won't be for Flatcar and was overwritten anyway and would be a duplicate
from the SERVER parameter in /usr/share/flatcar/update.conf.