os/sdk-modifying-flatcar.md: Document cork create usage
The default branches to be checked out will be the Alpha development
branches with the latest Alpha SDK.
See https://github.com/flatcar-linux/manifest/pull/14
The release branch of the manifest has changed from flatcar-build-x.y.z
to flatcar-CHANNEL-x.y.z which is updated here as well.
As replacement mechanism for the current CHANNEL-master.xml files
a new maintenance.xml is proposed that is part of the release tag/
branch of the manifest repository (not yet released).