Restructure the docs for them to show correctly in the new docs portal
We want to integrate the docs in the new Kinvolk docs portal, and that
requires making sure that the filesystem tree will generate a usable
docs structure (sidebar, sections, etc.) in the mentioned page.

So this patch reorganizes the folders and files, as a suggestion of a
structure.
The docs are also now all added into a "docs" folder, in order for extra
files (like the README) not to be included in the documentation
generation.