Document Terraform usage
A generic document describes how Terraform is used together with
Flatcar and Ignition. It covers the general principle. how to
generate the Ignition config and how to handle User Data changes.
An extra sections goes into the details of storing the User Data in
a remote location and rerunning Ignition in a safe way to circumvent
node recreation.
The provider-specific instructions for Terraform with Equinix Metal,
libvirt, and Hetzner are added to the cloud provider docs.