Add redirections from old addresses
This change adds redirections from the old addresses used in the old
Flatcar docs website, to the new addresses used here.

With this change, going to, for example:
  http://www.kinvolk.io/docs/flatcar-container-linux/latest/os/booting-on-azure/
Will redirect to:
  http://www.kinvolk.io/docs/flatcar-container-linux/latest/cloud-providers/booting-on-azure

The first pass was done automatically using the basename and assuming
everything was under the 'os' folder. After that, I manually fixed the
ignition folder (which existed before), and traced the _index.md files to
whatever they were called in the old site.

This should take care of all (or almost all) entries.