Rework Installing section
This change renames the "Quickstart" section to "Installing", as well as
moving "Cloud Providers", "Bare Metal" and "Community supported
platforms"as subsections of the new Installing section.

It also creates a new "Virtual Machines" subsection so that
virtualization technologies are not bunched together with Cloud, which
was weird.

Given the additional nesting, the files on disk have been renamed to
have shorter names, so that the URL paths are not so long (e.g.
cloud-providers/booting-on-azure is now installing/cloud/azure).

All aliases have been updated to make any existing links still work and
the links in the index file were also updated.