Rework the main index
The previous format used tables to try to cram as much information as
possible in the screen. This was already confusing before, but with the
new format, that uses much larger letters and has the left navigation
bar, the tables are really not useful at all. They are confusing to the
reader (because the columns are unrelated) and annoying to edit to the
editors. So, instead, make it a longer file that uses itemizations
instead of tables and groups the content in a way that makes more sense.

Also, reword the sections about "clusters". Most of those sections were
about managing Flatcar regardless of whether it was used for clusters or
not. So, only use clusters when they actually are clusters and use just
Flatcar for the rest.