Rework Setup and Operations section
We had a "Working with Clusters" section which was mostly about how to
configure Flatcar in general and very little about clusters in
particular. This led to users being confused when they were looking for
information that was not related to clusters. So, rename "Working with
Clusters" to "Setup and Operations" and then rename the sections inside
to be about what they should be.

This also re-organizes a lot of the files as many of them were in places
that made little sense.

Some links have been fixed, but as many pages are still switching
locations, I intend to do a full pass to review links after I'm done
with all relocations.