container-runtimes: simpler and future-proof containerd customization
The existing /usr/bin/containerd torcx wrapper already takes care of
using the right binaries. The torcx path in the customization
instructions can be avoided which helps to make this more future-proof
when torcx is not used. Also, the default config is under /etc and does
not need to be specified (the env var is custom to torcx, too).

Do not set up the env var nor specify the config path, and stop
hardcoding the torcx binary path.