Rework the update and rollback docs
The update and rollback docs had a lot of manual steps, some like the
restarting of update-engine were not really necessary.

Make use of the migration script that we maintain instead of outlining
manual steps to update from CoreOS. Also make use of the new
"flatcar-update" tool which allows to do an easy rollback to any older
version.