config-transpiler/dynamic-data.md: correct custom platform guide
The custom platform guide talks about using a drop-in unit but when a
non-OEM image is used, it doesn't have the service file at all.
The example used there and in the VMware docs was also lacking some
safety net for temporary malfunctioning metadata servers.

Instruct that the service file should be created instead of using a
drop-in (For existing OEM service files it will work as well to create
the service file under `/etc` which has a higher priority).
Also add some more options to have the oneshot unit be more robust.