Document how to avoid mismatching mounts after reboot
The kernel assigns /dev/sdX nodes by the order the disk is discovered
which can be different each boot. When a path such as /dev/sda is used
for mounting, this can result in a mismatch after reboot.
The better way is to always match the disk by the content that is
expected through the paths under /dev/disk/by-label/ and make sure the
labels get created on the first boot. An alternative when this is not
possible is to use /dev/disk/by-path/.