dev-guides: rework build order for testing temporary kconfig changes
We have constraints on the ordering so that the changes are fully
present in the image: pkg_setup phase of coreos-kernel needs to run
after coreos-modules has been merged, because config is fetched from
there and bootengine runs in pkg_setup.

This means: we should be experimenting on coreos-modules and need a full
rebuild of coreos-kernel at the end anyway.