Fully migrate Butane Config section
Rename the Config Transpiler section to Butane Config Transpiler to
differenciate it from the Container Linux Config section above.
Move out the operator-notes.md which is for CLC, and also move the
dynamic data docs to CLC and create a new one under the Ignition
section without using CLC features (Butane has no "magic" for that
which is actually not bad).