Consistent use of Butane vs Container Linux Config
While many parts were already converted to Butane, the text sometimes
used "Container Linux Config" and not all content was converted.
While checking I also noticed some small mistakes that were corrected.
The remaining use of Container Linux Config is mostly intentional, only
the Terraform examples are to be converted in the near future but since
they are self-contained, this is not urgent or inconsistent. The use
of CLC for GRUB configuration is also kept as alternative to the kernel
args setting in Butane when Ignition 2 is required.

Co-authored-by: Kai Lueke <kailuke@microsoft.com>